devop: refactor select network button
